feat(front): move directory-listing toggle into Authentication section
Address review feedback: relocate the 'List in workspace directory'
toggle from a standalone Discovery section into the Authentication
section (grouped with Invite by Link), and rewrite the description to be
explicit about the root multi-workspace sign-in domain and its link to
approved access domains. The domain is read from client config so it is
correct for self-hosted instances. Optimistically updates with rollback
on error, matching the sibling auth toggles.
